pyproject.toml
[project]
name = "llm-control-tutorial"
version = "0.1.0"
description = "Tutorial site for LLMs applied to control systems"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.7",
    "mkdocs-static-i18n>=1.2",
    "pymdown-extensions>=10.14",
    "mkdocs-glightbox>=0.4",
    "openai>=1.60",
    "python-dotenv>=1.0",
    "numpy>=2.0",
    "scipy>=1.14",
    "matplotlib>=3.9",
    "control>=0.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "jupytext>=1.16",
]
